# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Accounting Platform API
  version: 0.1.0
  description: >
    REST API for the AbhashTech multi-tenant accounting and bookkeeping
    platform with double-entry ledger, invoicing, expenses, payroll, tax,
    reporting, imports, backups, and investment accounting.
  contact:
    name: AbhashTech Accounting
    url: https://accounting.abhashtech.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.accounting.abhashtech.com/api/v1
    description: Production placeholder
...
security:
  - bearerAuth: []
paths:
  /bootstrap/first-admin:
    post:
      tags: [System]
      security: []
      summary: Create the first admin user and organization
      description: Only succeeds when no users exist in the system.
      operationId: bootstrapFirstAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapFirstAdminRequest'
      responses:
        '201':
          description: Bootstrap completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapFirstAdminResponse'
        default:
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Authenticate a user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        default:
          $ref: '#/components/responses/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh an access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        default:
          $ref: '#/components/responses/Error'
  /auth/password-reset/request:
    post:
      tags: [Auth]
      security: []
      summary: Request a password reset token
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
      responses:
        '200':
          description: Password reset requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestPasswordResetResponse'
        default:
          $ref: '#/components/responses/Error'
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      security: []
      summary: Confirm a password reset
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmPasswordResetRequest'
      responses:
        '200':
          description: Password reset confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmPasswordResetResponse'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/seed/india-defaults:
    post:
      tags: [Tax, Accounts, System]
      summary: Seed India default chart of accounts and GST presets
      operationId: seedIndiaDefaults
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Seed result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndiaSeedResult'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/closing/fiscal-years:
    get:
      tags: [Ledger]
      summary: List fiscal year closes
      operationId: listFiscalCloses
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Fiscal closes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FiscalClose'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Ledger]
      summary: Close a fiscal year
      operationId: closeFiscalYear
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseFiscalYearRequest'
      responses:
        '201':
          description: Fiscal year closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalClose'
        default:
          $ref: '#/components/responses/Error'
...
  /organizations/{organizationId}/data/export:
    get:
      tags: [System]
      summary: Export organization data as JSON
      description: Admin/Accountant-only tenant-scoped backup export containing accounting, tax, payroll, document metadata, import, budget, and investment records.
      operationId: exportOrganizationData
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Organization data export
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Suggested backup filename.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDataExport'
        default:
          $ref: '#/components/responses/Error'
  /organizations/{organizationId}/data/backups:
    get:
      tags: [System]
      summary: List backup snapshots
      operationId: listBackupSnapshots
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Backup snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BackupSnapshot'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [System]
      summary: Create a backup snapshot
      operationId: createBackupSnapshot
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBackupSnapshotRequest'
      responses:
        '201':
          description: Created backup snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupSnapshot'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: Request validation failed
            details:
              type: array
              items:
                type: object
                additionalProperties: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    BootstrapFirstAdminRequest:
      type: object
      required: [organization_name, admin_name, admin_email, admin_password]
      properties:
        organization_name:
          type: string
          example: Acme India
        admin_name:
          type: string
          example: Admin User
        admin_email:
          type: string
          format: email
        admin_password:
          type: string
          format: password
          minLength: 12
        base_currency:
          type: string
          example: INR
        country_code:
          type: string
          example: IN
        seed_india_defaults:
          type: boolean
          default: true
    BootstrapFirstAdminResponse:
      type: object
      properties:
        organization:
          $ref: '#/components/schemas/Organization'
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
            is_active:
              type: boolean
        membership:
          type: object
          properties:
            organization_id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            role:
              type: string
        india_seed:
          $ref: '#/components/schemas/IndiaSeedResult'
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    RequestPasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    RequestPasswordResetResponse:
      type: object
      required: [requested]
      properties:
        requested:
          type: boolean
        reset_token:
          type: string
          description: Development/staging reset token for local delivery until email transport is configured.
        reset_token_expires_at:
          type: string
          format: date-time
    ConfirmPasswordResetRequest:
      type: object
      required: [reset_token, new_password]
      properties:
        reset_token:
          type: string
        new_password:
          type: string
          minLength: 12
    ConfirmPasswordResetResponse:
      type: object
      required: [reset]
      properties:
        reset:
          type: boolean
    AuthTokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900
    Organization:
      type: object
      required: [id, name, base_currency, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        base_currency:
          type: string
          example: USD
        created_at:
          type: string
          format: date-time
...
    OrganizationDataExport:
      type: object
      required: [exported_at, organization]
      additionalProperties: true
      properties:
        exported_at:
          type: string
          format: date-time
        organization:
          $ref: '#/components/schemas/Organization'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        journal_transactions:
          type: array
          items:
            $ref: '#/components/schemas/JournalTransaction'
        audit_logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        invoices:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        expenses:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        tax_rates:
          type: array
          items:
            $ref: '#/components/schemas/TaxRate'
    BackupSnapshot:
      type: object
      required: [id, organization_id, file_name, storage_path, size_bytes, sha256, status]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        file_name:
          type: string
        storage_path:
          type: string
        size_bytes:
          type: integer
          format: int64
        sha256:
          type: string
        status:
          type: string
          example: completed
        completed_at:
          type: string
          format: date-time
          nullable: true
    CreateBackupSnapshotRequest:
      type: object
      properties:
        storage_path:
          type: string
          description: Optional local backup directory. Defaults to server backup storage configuration.
        retention_count:
          type: integer
          description: Optional count of newest snapshots to retain for this organization.
    Account:
      type: object
      required: [id, organization_id, name, code, type, is_active]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [asset, liability, equity, income, expense]
        subtype:
          type: string
          example: bank
        currency:
          type: string
          example: USD
        is_active:
          type: boolean
...
    JournalTransaction:
      type: object
      required: [id, organization_id, transaction_date, status, splits]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        transaction_date:
          type: string
          format: date
        memo:
          type: string
        source_module:
          type: string
          enum: [manual, invoice, credit_note, bill, expense, payroll, import, closing, payment]
        status:
          type: string
          enum: [draft, posted, reversed]
        splits:
          type: array
          items:
            $ref: '#/components/schemas/LedgerSplit'
    LedgerSplit:
      type: object
      required: [id, account_id, debit, credit]
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        memo:
          type: string
        debit:
          type: string
          deprecated: true
          example: '100.00'
        credit:
          type: string
          deprecated: true
          example: '0.00'
        debit_minor:
          type: integer
          format: int64
          example: 10000
        credit_minor:
          type: integer
          format: int64
          example: 0
        base_debit_minor:
          type: integer
          format: int64
          example: 835000
        base_credit_minor:
          type: integer
          format: int64
          example: 0
        currency:
          type: string
          example: USD
        exchange_rate_numerator:
          type: integer
          format: int64
          example: 8350
        exchange_rate_denominator:
          type: integer
          format: int64
          example: 100
        cleared:
          type: boolean
        reconciled:
          type: boolean
        reconciled_at:
          type: string
          format: date-time
          nullable: true
...
    Invoice:
      type: object
      required: [id, organization_id, invoice_number, status, subtotal_minor, tax_total_minor, total_minor]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        invoice_number:
          type: string
        issue_date:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, posted, paid, void]
        currency:
          type: string
          example: INR
        tax_inclusive:
          type: boolean
        subtotal_minor:
          type: integer
          format: int64
          example: 10000
        tax_total_minor:
          type: integer
          format: int64
          example: 1800
        total_minor:
          type: integer
          format: int64
          example: 11800
        accounts_receivable_id:
          type: string
          format: uuid
        pdf_attachment_id:
          type: string
          format: uuid
          nullable: true
        journal_transaction_id:
          type: string
          format: uuid
          nullable: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLine'
    InvoiceLine:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        quantity_millis:
          type: integer
          format: int64
          example: 1000
        unit_price_minor:
          type: integer
          format: int64
          example: 10000
        line_subtotal_minor:
          type: integer
          format: int64
        tax_amount_minor:
          type: integer
          format: int64
        line_total_minor:
          type: integer
          format: int64
        income_account_id:
          type: string
          format: uuid
        tax_rate_id:
          type: string
          format: uuid
          nullable: true
        tax_group_id:
          type: string
          format: uuid
          nullable: true
...
    Attachment:
      type: object
      required: [id, organization_id, file_name, storage_driver, storage_key, size_bytes]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        file_name:
          type: string
          example: INV-001.pdf
        content_type:
          type: string
          example: application/pdf
        storage_driver:
          type: string
          example: local
        storage_key:
          type: string
          example: invoices/INV-001.pdf
        size_bytes:
          type: integer
          format: int64
          example: 4096
...
    Expense:
      type: object
      required: [id, organization_id, expense_number, expense_date, status, subtotal_minor, tax_total_minor, total_minor]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        vendor_id:
          type: string
          format: uuid
          nullable: true
        expense_number:
          type: string
        expense_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, posted, void]
        currency:
          type: string
          example: INR
        tax_inclusive:
          type: boolean
        subtotal_minor:
          type: integer
          format: int64
          example: 10000
        tax_total_minor:
          type: integer
          format: int64
          example: 1800
        total_minor:
          type: integer
          format: int64
          example: 11800
        expense_account_id:
          type: string
          format: uuid
        payment_account_id:
          type: string
          format: uuid
        receipt_attachment_id:
          type: string
          format: uuid
          nullable: true
        journal_transaction_id:
          type: string
          format: uuid
          nullable: true
        reimbursable:
          type: boolean
...
    TaxRate:
      type: object
      required: [id, name, percentage, type, effective_from, is_active]
      properties:
        id:
          type: string
          format: uuid
        tax_authority_id:
          type: string
          format: uuid
        name:
          type: string
          example: GST 18%
        percentage:
          type: string
          example: '18.0000'
        percentage_basis:
          type: integer
          format: int64
          description: Percentage stored in ten-thousandths of one percent; 90000 means 9.0000%.
          example: 90000
        type:
          type: string
          enum: [VAT, GST, Sales Tax, Withholding]
        effective_from:
          type: string
          format: date
        effective_to:
          type: string
          format: date
          nullable: true
        is_compound:
          type: boolean
        is_active:
          type: boolean
...
    IndiaSeedResult:
      type: object
      required: [accounts_created, tax_rates_created, tax_groups_created, tax_authority_created]
      properties:
        accounts_created:
          type: integer
        tax_rates_created:
          type: integer
        tax_groups_created:
          type: integer
        tax_authority_created:
          type: boolean
...
    FiscalClose:
      type: object
      required: [id, organization_id, fiscal_year_start, fiscal_year_end, retained_earnings_account_id, net_income_minor, status, journal_transaction_id]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        fiscal_year_start:
          type: string
          format: date-time
        fiscal_year_end:
          type: string
          format: date-time
        retained_earnings_account_id:
          type: string
          format: uuid
        net_income_minor:
          type: integer
          format: int64
        status:
          type: string
          enum: [posted, reversed]
        journal_transaction_id:
          type: string
          format: uuid
    CloseFiscalYearRequest:
      type: object
      required: [fiscal_year_start, fiscal_year_end, retained_earnings_account_id]
      properties:
        fiscal_year_start:
          type: string
          format: date
        fiscal_year_end:
          type: string
          format: date
        retained_earnings_account_id:
          type: string
          format: uuid
    AuditLog:
      type: object
      required: [id, entity_type, entity_id, action, created_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        actor_user_id:
          type: string
          format: uuid
        entity_type:
          type: string
          example: invoice
        entity_id:
          type: string
          format: uuid
        action:
          type: string
          example: post
        before_json:
          type: string
        after_json:
          type: string
        ip_address:
          type: string
        user_agent:
          type: string
        created_at:
          type: string
          format: date-time
...